backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
pydantic
python-dotenv
langchain-openai
langchain-ollama
langchain_community
supabase


gotrue
langgraph-checkpoint-postgres
psycopg-binary
psycopg-pool
